Revert "Improved duration precision to decimal"
This reverts commit 4e0349b9da2dacc6d81e5f7418f4a08dd9e0c86d.

+++ test/MetadataUtility.Tests/TestHelpers/FixtureModel.cs
@@ -36,7 +36,7 @@ namespace MetadataUtility.Tests.TestHelpers
 
         public string MimeType { get; set; }
 
-        public decimal DurationSeconds { get; set; }
+        public double DurationSeconds { get; set; }
 
         public byte Channels { get; set; }
